# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Research Agent API",
    "description": "Production-grade REST API for the Research Agent platform. Supports academic literature search, document ingestion, retrieval-augmented generation (RAG) question answering, summarization, and similarity analysis.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search academic literature",
        "description": "Search across configured academic literature databases.",
        "operationId": "search_articles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Search, download, process, and index articles",
        "description": "Runs the full ingest pipeline: search \u2192 download \u2192 process \u2192 index.",
        "operationId": "ingest_articles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ask": {
      "post": {
        "tags": [
          "ask"
        ],
        "summary": "Ask a research question",
        "description": "Query indexed documents with a natural language question. Uses RAG (retrieval-augmented generation) to answer from your document corpus.",
        "operationId": "ask_question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "409": {
            "description": "Agent is busy processing another request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer",
            "default": ""
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model",
            "default": ""
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider",
            "default": ""
          },
          "llm_elapsed_ms": {
            "type": "number",
            "title": "Llm Elapsed Ms",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "AnswerResponse"
      },
      "ArticleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Authors"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "abstract": {
            "type": "string",
            "title": "Abstract",
            "default": ""
          }
        },
        "type": "object",
        "title": "ArticleResponse"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question",
            "description": "Research question"
          },
          "llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider",
            "description": "LLM provider override"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "LLM temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "Max tokens in response"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AnswerResponse"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkInfo"
            },
            "type": "array",
            "title": "Chunks"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AskResponse"
      },
      "ChunkInfo": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "heading": {
            "type": "string",
            "title": "Heading",
            "default": ""
          }
        },
        "type": "object",
        "title": "ChunkInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestDocumentResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks",
            "default": 0
          }
        },
        "type": "object",
        "title": "IngestDocumentResponse"
      },
      "IngestRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Search and ingest query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum articles to ingest",
            "default": 5
          },
          "download_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Dir",
            "description": "Custom download directory"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "IngestRequest"
      },
      "IngestResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestDocumentResponse"
            },
            "type": "array",
            "title": "Documents"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "IngestResponse"
      },
...
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Search query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum articles to return",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleResponse"
            },
            "type": "array",
            "title": "Articles"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "SearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}